ft_strcat

+++ ft_strlen.c
@@ -2,13 +2,15 @@
  * File              : ft_strlen.c
  * Author            : Mikael Berglund <mikael.berglund2@gmail.com>
  * Date              : 20.10.2019
- * Last Modified Date: 20.10.2019
+ * Last Modified Date: 23.10.2019
  * Last Modified By  : Mikael Berglund <mikael.berglund2@gmail.com>
  */
 
-unsigned long	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
-	unsigned long len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
